Added implementation for stone placement status
Added implementation for stone placement status. moveMade method to
change all the stones' status to old (not moved this turn). Must be
called before placement of new stone. Changed player makeMove method to
reflect this. Added status class field to Stone to hold this status and
the setter and getter methods for it.